DatabaseEditor/src/ru/ucoz/ctor/servlets/CancelEditing.java
package ru.ucoz.ctor.servlets;

import java.io.IOException;
import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

@WebServlet("/cancel")
public class CancelEditing extends HttpServlet {
	private static final long serialVersionUID = 1L;

	protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
		request.removeAttribute("add");
		request.removeAttribute("save");
		request.getRequestDispatcher("./").forward(request, response);
	}
}
